feat(writing): enhance reflections page with improved UX
Improvements:
- Changed date format from "October" to "Oct" (abbreviated months)
- Added clock icon before reading time using lucide-react
- Added | separator between date and read time
- Made date and read time smaller (text-xs) and lighter (text-text-secondary)
- Converted tags to pill badges using badge classes
- Highlighted latest post in separate featured section with card styling
- Split posts into "Latest" and "All Posts" sections
- Updated hover states with accent-warm color
- Added gradient dividers between sections

Posts are already sorted in descending order (newest first).
Reading time is calculated at 200 words per minute.

+++ lib/utils.ts
@@ -21,7 +21,7 @@ export function formatDate(dateString: string): string {
   const date = new Date(dateString)
   return date.toLocaleDateString('en-US', {
     year: 'numeric',
-    month: 'long',
+    month: 'short',
     day: 'numeric',
   })
 }
@@ -29,10 +29,7 @@ export function formatDate(dateString: string): string {
 /**
  * Format a date range (start - end)
  */
-export function formatDateRange(
-  startDate: string,
-  endDate?: string | null
-): string {
+export function formatDateRange(startDate: string, endDate?: string | null): string {
   const start = new Date(startDate)
   const startFormatted = start.toLocaleDateString('en-US', {
     year: 'numeric',
@@ -55,16 +52,12 @@ export function formatDateRange(
 /**
  * Calculate duration between two dates in months
  */
-export function calculateDuration(
-  startDate: string,
-  endDate?: string | null
-): number {
+export function calculateDuration(startDate: string, endDate?: string | null): number {
   const start = new Date(startDate)
   const end = endDate ? new Date(endDate) : new Date()
 
   const months =
-    (end.getFullYear() - start.getFullYear()) * 12 +
-    (end.getMonth() - start.getMonth())
+    (end.getFullYear() - start.getFullYear()) * 12 + (end.getMonth() - start.getMonth())
 
   return months
 }
